+++ DaemonMC/Network/Bedrock/StartGame.cs
@@ -32,7 +32,7 @@ namespace DaemonMC.Network.Bedrock
         protected override void Encode(PacketEncoder encoder)
         {
             encoder.WriteSignedVarLong(EntityId);
-            encoder.WriteVarLong((ulong) EntityId);
+            encoder.WriteVarLong(EntityId);
             encoder.WriteVarInt(GameMode);
             encoder.WriteVec3(Position);
             encoder.WriteVec2(Rotation);
